+++ tests/testthat/test-export_rasters.R
@@ -14,7 +14,7 @@ test_that("export_rasters() works with GeoTIFFs", {
   # Expected files
   files <- sort(list.files(dir))
   n <- length(files)
-  expect_snapshot(cat(files[c(1:8, (n-4):n)], sep = "\n"))
+  expect_snapshot(cat(files[c(1:8, (n - 4):n)], sep = "\n"))
 
 
   # First exported distribution is very close to first distribution
@@ -47,7 +47,7 @@ test_that("export_rasters() works with PNG and reprojection", {
   # Expected files
   png_export_files <- sort(list.files(dir))
   n <- length(png_export_files)
-  expect_snapshot(cat(png_export_files[c(1:8, (n-4):n)], sep = "\n"))
+  expect_snapshot(cat(png_export_files[c(1:8, (n - 4):n)], sep = "\n"))
 
   # First exported distribution is very close to first distribution
   file <- file.path(dir, png_export_files[grep("distr", png_export_files)[1]])